type : refactor MVC 패턴에 맞게 재분류
body :
