+++ src/main/java/bridge/model/BridgeRandomNumberGenerator.java
@@ -1,5 +1,6 @@
-package bridge;
+package bridge.model;
 
+import bridge.model.BridgeNumberGenerator;
 import camp.nextstep.edu.missionutils.Randoms;
 
 public class BridgeRandomNumberGenerator implements BridgeNumberGenerator {
