+++ src/main/java/bridge/model/CompareBridge.java
@@ -1,0 +1,4 @@
+package bridge.model;
+
+public class CompareBridge {
+}
