+++ src/main/java/bridge/validator/Validator.java
@@ -1,0 +1,4 @@
+package bridge.validator;
+
+public class Validator {
+}
